Finished, or as near as damn it.
The FPS seems to be jittery, but I can't decide if it's just the difference in colour that makes it more pronounced or not.

I've renamed the package to that of the original SunBeam LWP, added in the original resources, and mesh. I also removed some junk data from the mesh. For isntnace, 0.500 as a float is pointless. The trailing 0's are lost anyway.